+++ src/validation/decorators/index.ts
@@ -8,6 +8,7 @@ export {
   HasUniqueValues,
   hasUniqueValues,
 } from './has-unique-values.decorator.js';
+export { IsDateType } from './is-date-type.decorator.js';
 export {
   IsGreaterThanProperty,
   isGreaterThanValue,
@@ -19,3 +20,4 @@ export {
 } from './is-object-with-keys-max-length.decorator.js';
 export { RequiresPropertyNotToBeDefined } from './requires-property-not-to-be-defined.decorator.js';
 export { RequiresProperty } from './requires-property.decorator.js';
+export { ValidateNestedType } from './validate-nested-type.decorator.js';